Apply massimomorara's fixes to 0.98.4 release

+++ src/jsoncons/json_input_handler.hpp
@@ -258,7 +258,7 @@ private:
     {
     }
 
-    void do_integer_value(long long, const basic_parsing_context<CharT>&) override
+    void do_integer_value(int64_t, const basic_parsing_context<CharT>&) override
     {
     }
